Pull up RemoveStatsOverride
